Refactor pageStructure component
